refactor: split executor.rs into submodules by responsibility
Split the 3965-line executor.rs into focused submodules:

- basic.rs (360 lines): Single-effect execution, resource resolution,
  Secret unwrapping, binding map updates
- parallel.rs (437 lines): Dependency computation, fine-grained
  parallel scheduling
- replace.rs (384 lines): Replace effect orchestration (CBD/DBD)
- phased.rs (1158 lines): Interdependent Replace ordering (4-phase)
- mod.rs: Public types, execute_plan entry point, tests

No logic changes - pure structural refactoring. All 23 executor
tests pass unchanged.

Refs #1850
